fix(boot): expand packaged kernel capacity

+++ Makefile
@@ -3,12 +3,16 @@ BOCHS := bochs -q
 DISK = hd.img
 INIT_BOOT_CODE = MBR.bin
 LOADER = loader.img
-CORE   = core.img
-CORESYM   = $(BUILD_DIR)/core_symbol.img
+CORE   = core/build/core.img
+CORESYM   = core/build/core_symbol.img
 FONT   = hankaku_font.img
 TEST_PROC = core/apps/build/compositor
 # TEST_PROC = core/apps/build/ls
 TEST_IMG= core/apps/test/b.bmp
+SECTOR_SIZE := 512
+LOADER_SECTOR_COUNT := 11
+CORE_START_SECTOR := 13
+CORE_SECTOR_COUNT := 512
 
 # Use ELF format
 #kernel code ###########################
@@ -82,15 +86,39 @@ init_boot_code: $(INIT_BOOT_CODE)
 # 4. Burn test things (programs)
 # 5. Burn test things (images)
 mount: init_boot_code loader.img core.img font
-	dd if=$(LOADER) of=$(DISK) bs=512 count=300 seek=2 conv=notrunc #loader
-	dd if=$(CORE) of=$(DISK) bs=512 count=300 seek=13 conv=notrunc  #core 122k (blank is 1M)
+	@loader_size=$$(wc -c < "$(LOADER)"); \
+	loader_limit=$$(($(LOADER_SECTOR_COUNT) * $(SECTOR_SIZE))); \
+	if [ "$$loader_size" -gt "$$loader_limit" ]; then \
+		printf '%s is %s bytes; MBR loader limit is %s\n' \
+			"$(LOADER)" "$$loader_size" "$$loader_limit" >&2; exit 1; \
+	fi
+	@core_size=$$(wc -c < "$(CORE)"); \
+	core_limit=$$(($(CORE_SECTOR_COUNT) * $(SECTOR_SIZE))); \
+	if [ "$$core_size" -gt "$$core_limit" ]; then \
+		printf '%s is %s bytes; boot image limit is %s\n' \
+			"$(CORE)" "$$core_size" "$$core_limit" >&2; exit 1; \
+	fi
+	dd if=$(LOADER) of=$(DISK) bs=$(SECTOR_SIZE) count=$(LOADER_SECTOR_COUNT) seek=2 conv=notrunc #loader
+	dd if=$(CORE) of=$(DISK) bs=$(SECTOR_SIZE) count=$(CORE_SECTOR_COUNT) seek=$(CORE_START_SECTOR) conv=notrunc #core, reserved through sector 524
 	dd if=$(FONT) of=$(DISK) bs=512 count=300 seek=2048 conv=notrunc #font.img for now size 4k place to offset 1M
 	dd if=$(TEST_PROC) of=$(DISK) bs=512 count=300 seek=3000 conv=notrunc
 	dd if=$(TEST_IMG) of=$(DISK) bs=512 count=300 seek=6144 conv=notrunc # place to 3M img size < 150k
 
 mount_debug: init_boot_code loader.img core_symbol.img font
-	dd if=$(LOADER) of=$(DISK) bs=512 count=300 seek=2 conv=notrunc #loader
-	dd if=$(CORESYM) of=$(DISK) bs=512 count=300 seek=13 conv=notrunc  #core 122k (blank is 1M)
+	@loader_size=$$(wc -c < "$(LOADER)"); \
+	loader_limit=$$(($(LOADER_SECTOR_COUNT) * $(SECTOR_SIZE))); \
+	if [ "$$loader_size" -gt "$$loader_limit" ]; then \
+		printf '%s is %s bytes; MBR loader limit is %s\n' \
+			"$(LOADER)" "$$loader_size" "$$loader_limit" >&2; exit 1; \
+	fi
+	@core_size=$$(wc -c < "$(CORESYM)"); \
+	core_limit=$$(($(CORE_SECTOR_COUNT) * $(SECTOR_SIZE))); \
+	if [ "$$core_size" -gt "$$core_limit" ]; then \
+		printf '%s is %s bytes; boot image limit is %s\n' \
+			"$(CORESYM)" "$$core_size" "$$core_limit" >&2; exit 1; \
+	fi
+	dd if=$(LOADER) of=$(DISK) bs=$(SECTOR_SIZE) count=$(LOADER_SECTOR_COUNT) seek=2 conv=notrunc #loader
+	dd if=$(CORESYM) of=$(DISK) bs=$(SECTOR_SIZE) count=$(CORE_SECTOR_COUNT) seek=$(CORE_START_SECTOR) conv=notrunc #core, reserved through sector 524
 	dd if=$(FONT) of=$(DISK) bs=512 count=300 seek=2048 conv=notrunc #font.img for now size 4k
 	# dd if=$(TEST_PROC) of=$(DISK) bs=512 count=300 seek=3000 conv=notrunc
 	# dd if=$(TEST_IMG) of=$(DISK) bs=512 count=300 seek=6144 conv=notrunc # place at 3M, img size < 150k
